Got it running...sort of
